Commit for release 0.3.0-pre3
I'm rushing this release too because I want to install a Lubuntu dual
boot today, and any number of things could go wrong between now and next
week(or even tomorrw, the scheduled release date).
This commit includes all the cleanup necessary to make the game playable
and the code nice-looking. I disabled a mixin that replaces the player
inventory because it's not finished yet.